cleanup for undo-jsp-precompile.patch
the original "Pre-compile the JSP pages" commit involved a bunch of whitespace
formatting changes, which included lines like the gums version numbers.

That made reverting the change (for el5) more difficult, because any future
changes to the touched files (the pom.xml's) wouldn't apply after reverting
the JSP precomile commit (because of the whitespace changes).

I'm cleaning up this patch to remove only the non-whitespace changes that were
added in the original JSP precomile commit.  This should make it so we don't
need to redo this patch every time the gums version changes, and it should also
make it easier to update if later pom.xml changes make that necessary.

https://github.com/edquist/gums/compare/software-1654-undo-jsp-precompile


git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/trunk@20452 4e558342-562e-0410-864c-e07659590f8c