Working build for el6 and el7
Used 'maven' instead of 'maven-local' for the package, and 'mvn' instead
of 'xmvn' for the executable for el7.  Also removed OS libraries for the
maven plugins, using the ones requested by the project's pom.xml
instead.

By setting the "maven_offline" macro in the spec file, we can use a
pre-downloaded set of maven dependencies to do the build (which are
located in the upstream area), instead of downloading them live.

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/osg-3.3@21449 4e558342-562e-0410-864c-e07659590f8c