revert SOFTWARE-2075 changes that were based on gratia 1.16.3
because the gratia-service-1.16.3-2 builds appear to be breaking the VMU
tests, and all of my changes on top of it also don't work totally.

At some point we can investigate what in gratia 1.16.3 actually breaks
things.  I'm only reverting to what Kevin had in development, and the
rest of these changes are moving to branches/software-2075/gratia, which
is based on 1.16.2.

svn merge -c-21918 -c-21842 -c-21835 -c-21834 -c-21833 -c-21832 .


git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/trunk@21938 4e558342-562e-0410-864c-e07659590f8c