drop our (unused) copy of do-component-build (SOFTWARE-2161)
the one that actually gets used (and patched) is the one from
hadoop-bigtop-packaging.tar.gz, in the src.rpm.  It's misleading because
the src.rpm also contains a do-component-build source file, which is
identical to the one in bigtop-packaging, but the latter is apparently
the one that gets used--not the former, which ours overrides.
So, keeping ours is just misleading.

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/software-2161@22539 4e558342-562e-0410-864c-e07659590f8c