build fix: drop mvn304.patch and tweak PATH instead (SOFTWARE-3280)
The mvn304.patch was not applying because the do-component-build script
has apparently changed significantly enough.  All that script did was
hard-code the path for the mvn command, which we can address outside of
the script by just prepending it to PATH.  This is convenient since the
location and number of times where mvn is called in do-component-build
is apparently subject to change, and just tweaking PATH addresses all of
them.

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/upcoming@24475 4e558342-562e-0410-864c-e07659590f8c