patch out missing stuff from packaging_functions.sh (SOFTWARE-3280)
Cloudera apparently refers to a common.sh (which it does not package) that this
script expects to find under /, which needs to be set in the build
environment.  Not very nice.  Hopefully print_linkinfo() is the only thing from it
that got used, which doesn't do anything important anyway.

It's tricky because we are patching a Source file directly, rather than patching
some file that got extracted from the Source0 tarball.  So we don't use the %patch0
macro here.

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/upcoming@24477 4e558342-562e-0410-864c-e07659590f8c