pull in apache-tomcat tarball source (SOFTWARE-4540)
turns out the hadoop build downloads this archive at build time!

also, skip the apache-tomcat download.  it looks like it's supposed
to skip the download anyway if it's already present, but that did not
seem to work for me, so i've commented out that logic.  Instead the
spec %prep section will copy it into place.

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/upcoming@26347 4e558342-562e-0410-864c-e07659590f8c