introduce osg-next repos (SOFTWARE-4509)
This adds osg-next yum repo definitions that point to the next OSG
release series.

The idea is to be able to run:

  yum update --enablerepo=osg-next osg-release

With, in this case, osg-release 3.6 installed, and it will upgrade to
the osg-release 3.7 from the osg-3.7-release repos.

It's possible to pull in arbitrary packages from osg-next, but it is
perhaps most useful for the osg-release package itself, as otherwise the
only way to upgrade is to give a full url to the rpm.

The other osg-next-* repos are also defined, in case you want to update
to the next release series (eg, OSG 3.7) before the 'osg-release' rpm
for the next OSG series has been officially released.

As in:

  yum update --enablerepo=osg-next-development osg-release

git-svn-id: https://vdt.cs.wisc.edu/svn/native/redhat/branches/osg-3.6@26964 4e558342-562e-0410-864c-e07659590f8c